Renames filter directives
