+++ webapp/src/js/directives/date-filter.js
