+++ webapp/src/js/directives/facility-filter.js
