+++ webapp/src/js/directives/freetext-filter.js
